requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
appdirs=1.4.4=py_0
appnope=0.1.2=py37hecd8cb5_1001
argon2-cffi=20.1.0=py37h9ed2024_1
async_generator=1.10=py37h28b3542_0
attrs=21.2.0=pyhd3eb1b0_0
audioread=2.1.9=py37hf985489_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bleach=4.0.0=pyhd3eb1b0_0
brotlipy=0.7.0=py37h9ed2024_1003
bzip2=1.0.8=h1de35cc_0
ca-certificates=2021.9.30=hecd8cb5_1
certifi=2021.5.30=py37hecd8cb5_0
cffi=1.14.6=py37h2125817_0
chardet=4.0.0=py37hecd8cb5_1003
cryptography=3.4.7=py37h2fd3fbb_0
cycler=0.10.0=py37_0
cython=0.29.24=py37h23ab428_0
dbus=1.13.18=h18a8e69_0
decorator=5.0.9=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
docopt=0.6.2=py37hecd8cb5_0
entrypoints=0.3=py37_0
et_xmlfile=1.1.0=py37hecd8cb5_0
expat=2.4.1=h23ab428_2
ffmpeg=4.2.2=h97e5cf8_0
freetype=2.10.4=ha233b18_0
future=0.18.2=py37_1
gettext=0.21.0=h7535e17_0
glib=2.69.0=hdf23fa2_0
gmp=6.2.1=h23ab428_2
gnutls=3.6.15=hed9c0bf_0
icu=58.2=h0a44026_3
idna=2.10=pyhd3eb1b0_0
importlib-metadata=3.10.0=py37hecd8cb5_0
importlib_metadata=3.10.0=hd3eb1b0_0
intel-openmp=2021.3.0=hecd8cb5_3375
ipykernel=5.3.4=py37h5ca1d4c_0
ipython=7.26.0=py37h01d92e1_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.3=pyhd3eb1b0_1
jdcal=1.4.1=py_0
jedi=0.18.0=py37hecd8cb5_1
jinja2=3.0.1=pyhd3eb1b0_0
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9b=he5867d9_2
jsonpickle=2.0.0=pyhd3eb1b0_0
jsonschema=3.2.0=py_2
jupyter=1.0.0=py37_7
jupyter_client=6.1.12=pyhd3eb1b0_0
jupyter_console=6.4.0=pyhd3eb1b0_0
jupyter_core=4.7.1=py37hecd8cb5_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kiwisolver=1.3.1=py37h23ab428_0
lame=3.100=h1de35cc_0
lcms2=2.12=hf1fd2bf_0
libcxx=10.0.0=1
libffi=3.3=hb1e8313_2
libflac=1.3.3=h4a8c4bd_0
libgfortran=3.0.1=h93005f0_2
libiconv=1.16=h1de35cc_0
libidn2=2.3.2=h9ed2024_0
libllvm10=10.0.1=h76017ad_5
libogg=1.3.5=h9ed2024_1
libopus=1.3.1=h1de35cc_0
libpng=1.6.37=ha441bb4_0
libprotobuf=3.17.2=h2842e9f_1
librosa=0.8.1=pyhd8ed1ab_0
libsndfile=1.0.28=h4a8c4bd_1000
libsodium=1.0.18=h1de35cc_0
libtasn1=4.16.0=h9ed2024_0
libtiff=4.2.0=h87d7836_0
libunistring=0.9.10=h9ed2024_0
libuv=1.40.0=haf1e3a3_0
libvorbis=1.3.7=haf1e3a3_0
libvpx=1.7.0=h378b8a2_0
libwebp-base=1.2.0=h9ed2024_0
libxml2=2.9.12=hcdb78fc_0
llvm-openmp=10.0.0=h28b9765_0
llvmlite=0.36.0=py37he4411ff_4
lz4-c=1.9.3=h23ab428_1
madmom=0.16.1=pypi_0
markupsafe=2.0.1=py37h9ed2024_0
matplotlib=3.3.4=py37hecd8cb5_0
matplotlib-base=3.3.4=py37h8b3ea08_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mido=1.2.10=pypi_0
mir_eval=0.6=pyh9f0ad1d_0
mistune=0.8.4=py37h1de35cc_0
mkl=2021.3.0=hecd8cb5_517
mkl-service=2.4.0=py37h9ed2024_0
mkl_fft=1.3.0=py37h4a7008c_2
mkl_random=1.2.2=py37hb2f4e1b_0
munch=2.5.0=pyhd3eb1b0_0
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.1.0=py37hecd8cb5_0
nbformat=5.1.3=pyhd3eb1b0_0
ncurses=6.2=h0a44026_1
nest-asyncio=1.5.1=pyhd3eb1b0_0
nettle=3.7.3=h230ac6f_1
ninja=1.10.2=hf7b0b51_1
notebook=6.4.2=py37hecd8cb5_0
numba=0.53.0=py37hb2f4e1b_0
numpy=1.20.3=py37h4b4dc7a_0
numpy-base=1.20.3=py37he0bd621_0
olefile=0.46=py37_0
openh264=2.1.0=hd9629dc_0
openjpeg=2.3.0=hb95cd4c_1
openpyxl=3.0.7=pyhd3eb1b0_0
openssl=1.1.1l=h9ed2024_0
packaging=21.0=pyhd3eb1b0_0
pandas=1.2.5=py37h23ab428_0
pandocfilters=1.4.3=py37hecd8cb5_1
parso=0.8.2=pyhd3eb1b0_0
pcre=8.45=h23ab428_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.3.1=py37ha4cf6ea_0
pip=21.1.3=py37hecd8cb5_0
pooch=1.4.0=pyhd3eb1b0_0
pretty-midi=0.2.9=pypi_0
prometheus_client=0.11.0=pyhd3eb1b0_0
prompt-toolkit=3.0.17=pyh06a4308_0
prompt_toolkit=3.0.17=hd3eb1b0_0
protobuf=3.17.2=py37h23ab428_0
ptyprocess=0.7.0=pyhd3eb1b0_2
py-cpuinfo=8.0.0=pyhd3eb1b0_1
pycparser=2.20=py_2
pygments=2.9.0=pyhd3eb1b0_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py37h655552a_2
pyrsistent=0.17.3=py37haf1e3a3_0
pysocks=1.7.1=py37hecd8cb5_0
pysoundfile=0.10.3.post1=pyhd3deb0d_0
python=3.7.10=h88f2d9e_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.7=2_cp37m
pytorch=1.9.0=py3.7_0
pytz=2021.1=pyhd3eb1b0_0
pyzmq=22.2.1=py37h23ab428_1
qt=5.9.7=h468cd18_1
qtconsole=5.1.0=pyhd3eb1b0_0
qtpy=1.9.0=py_0
readline=8.1=h9ed2024_0
requests=2.25.1=pyhd3eb1b0_0
resampy=0.2.2=py_0
sacred=0.7.4=py_0
scikit-learn=0.24.2=py37hb2f4e1b_0
scipy=1.6.2=py37hd5f7400_1
send2trash=1.5.0=pyhd3eb1b0_1
setuptools=52.0.0=py37hecd8cb5_0
sip=4.19.8=py37h0a44026_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.36.0=hce871da_0
tensorboardx=2.4=pyhd8ed1ab_0
terminado=0.9.4=py37hecd8cb5_0
testpath=0.5.0=pyhd3eb1b0_0
threadpoolctl=2.2.0=pyhbf3da8f_0
tk=8.6.10=hb0a8c7a_0
torchaudio=0.9.0=py37
torchvision=0.10.0=py37_cpu
tornado=6.1=py37h9ed2024_0
tqdm=4.62.1=pyhd3eb1b0_1
traitlets=5.0.5=pyhd3eb1b0_0
typing_extensions=3.10.0.0=pyh06a4308_0
urllib3=1.26.6=pyhd3eb1b0_1
wcwidth=0.2.5=py_0
webencodings=0.5.1=py37_1
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.5.1=py37_0
wrapt=1.12.1=py37h1de35cc_1
x264=1!157.20191217=h1de35cc_0
xlrd=2.0.1=pyhd3eb1b0_0
xz=5.2.5=h1de35cc_0
zeromq=4.3.4=h23ab428_0
zipp=3.5.0=pyhd3eb1b0_0
zlib=1.2.11=h1de35cc_3
zstd=1.4.9=h322a384_0
